requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.5.6
autopep8==1.5.7
boto3==1.16.51
botocore==1.19.51
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.8
django-admin==2.0.1
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==3.4.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.3.0
django-jsonfield==1.4.1
django-mysql==3.11.1
django-six==1.0.4
django-storages==1.11.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Faker==5.0.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
mailjet-rest==1.3.0
Markdown==3.2.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pep8==1.7.1
Pillow==7.2.0
progress==1.5
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.8.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.1
pytoml==0.1.21
pytz==2018.9
razorpay==1.2.0
rcssmin==1.0.6
requests==2.24.0
retrying==1.3.3
rjsmin==1.1.0
s3transfer==0.3.3
screen==1.0.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.10
webencodings==0.5.1
wrapt==1.12.1
xlwt==1.3.0
